Made search better + nice index.html + fixed bugs
After the change, the search algorithm doesn't check for exact phrases and words. For example if you input "auth", it will still understand that you want to search something that includes "authentication" as its data.

Also the search box is auto selected when the page is loaded.

The index.html now has a dark background and a loading text, that will self enable if the app is loading for more than 0.75 second

Fixed bugs:
 - The Update button doesn't get enabled after changing textbox
- If the user is not logged in, the "Save Snippet" text does not properly change to "Log In"